refactor(cozy-harvest-lib): Extract ReadOnlyIdentifier
