+++ packages/cozy-harvest-lib/src/components/AccountForm/index.jsx
@@ -2,46 +2,20 @@ import React, { PureComponent } from 'react'
 import { Form } from 'react-final-form'
 import PropTypes from 'prop-types'
 import get from 'lodash/get'
-import cx from 'classnames'
 
 import { isMobile } from 'cozy-device-helper'
 import Button from 'cozy-ui/transpiled/react/Button'
-import Card from 'cozy-ui/transpiled/react/Card'
-import Icon from 'cozy-ui/transpiled/react/Icon'
-import { Media, Img, Bd } from 'cozy-ui/transpiled/react/Media'
 import withLocales from '../hoc/withLocales'
 
 import AccountFields from './AccountFields'
+import ReadOnlyIdentifier from './ReadOnlyIdentifier'
 import TriggerErrorInfo from '../infos/TriggerErrorInfo'
 import { getEncryptedFieldName } from '../../helpers/fields'
 import { KonnectorJobError } from '../../helpers/konnectors'
 import manifest from '../../helpers/manifest'
 import withKonnectorLocales from '../hoc/withKonnectorLocales'
-import KonnectorIcon from '../KonnectorIcon'
 
 const VALIDATION_ERROR_REQUIRED_FIELD = 'VALIDATION_ERROR_REQUIRED_FIELD'
-
-const ReadOnlyIdentifier = props => {
-  const { className, onClick, konnector, identifier, ...rest } = props
-
-  return (
-    <Card
-      className={cx({ 'u-c-pointer': onClick }, className)}
-      onClick={onClick}
-      {...rest}
-    >
-      <Media>
-        <Img>
-          <KonnectorIcon konnector={konnector} className="u-w-1-half u-mr-1" />
-        </Img>
-        <Bd>{identifier}</Bd>
-        <Img>
-          <Icon icon="bottom-select" />
-        </Img>
-      </Media>
-    </Card>
-  )
-}
 
 /**
  * AccountForm is reponsible of generating a form which will allow user to
